Move database helper methods into separate class
